% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county}
\alias{county}
\title{SF object with county borders in Sweden}
\format{
SF object with 3 columns
\describe{
  \item{ln_kod}{a character with county ID}
  \item{ln_namn}{a character with county name}
  \item{geometry}{a multipolygon with county borders}
}
}
\source{
{SCB, https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/}
}
\usage{
county
}
\description{
Includes polygon and regional IDs for counties in Sweden
}
\keyword{datasets}
