% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fa_region}
\alias{fa_region}
\title{SF object with FA region borders in Sweden}
\format{
SF object with 3 columns
\describe{
  \item{fa_kod}{a character with FA region ID}
  \item{fa_namn}{a character with FA region name}
  \item{geometry}{a multipolygon with FA region borders}
}
}
\source{
{SCB, https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/}
}
\usage{
fa_region
}
\description{
Includes polygon and regional IDs for FA regions in Sweden
}
\keyword{datasets}
