% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbg}
\alias{gbg}
\title{SF object with borders in Gothenburg}
\format{
SF object with 4 columns
\describe{
  \item{ajour_datu}{a date describing last update}
  \item{basomrade}{a character describing area code}
  \item{area_km2}{a number describing area size in km2}
  \item{geometry}{a multipolygon with region borders}
}
}
\source{
{Stockholm stad, https://dataportalen.stockholm.se}
}
\usage{
gbg
}
\description{
Includes polygon and metadata about Gothenburg
}
\keyword{datasets}
