% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malmo}
\alias{malmo}
\title{SF object with borders in Malmo}
\format{
SF object with 4 columns
\describe{
  \item{delomrade}{a character with region name}
  \item{do_kod}{a number with id code?}
  \item{area}{a number describing area size}
  \item{geometry}{a multipolygon with region borders}
}
}
\source{
{Stockholm stad, https://dataportalen.stockholm.se}
}
\usage{
malmo
}
\description{
Includes polygon and metadata about Malmo
}
\keyword{datasets}
