% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_so.R
\name{\%>\%}
\alias{\%>\%}
\alias{load_deso}
\alias{load_regso}
\title{Load DeSO polgyons}
\usage{
load_deso()

load_regso()
}
\description{
Function to read Swedish Demografiska statistikomraden (DeSO)

Function to read Swedish regionala statistikomrden (RegSO)
}
