% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sthlm}
\alias{sthlm}
\title{SF object with borders in Stockholm}
\format{
SF object with 62 columns
\describe{
  \item{namn}{a character with region name}
  \item{nyckelkod}{a number with key}
  \item{geometry}{a multipolygon with region borders}
}
}
\source{
{Stockholm stad, https://dataportalen.stockholm.se}
}
\usage{
sthlm
}
\description{
Includes polygon and metadata about Stockholm
}
\keyword{datasets}
