% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sthlm_points}
\alias{sthlm_points}
\title{SF object with region points in Stockholm}
\format{
SF object with 66 columns
\describe{
  \item{namn}{a character with region name}
  \item{nyckelkod}{a number with key}
  \item{geometry}{a multipolygon with region borders}
}
}
\source{
{Stockholm stad, https://dataportalen.stockholm.se}
}
\usage{
sthlm_points
}
\description{
Includes st points and metadata about Stockholm
}
\keyword{datasets}
