% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_swemap.R
\name{theme_swemap}
\alias{theme_swemap}
\title{Clean theme for maps}
\usage{
theme_swemap(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
A clean theme that is good for displaying maps from swemaps2
}
