% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_swemap.R
\name{theme_swemap2}
\alias{theme_swemap2}
\title{Clean theme with gridlines for maps}
\usage{
theme_swemap2(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
A clean theme for maps in ggplot2, but still with gridlines
}
